MAINT: move functions.py into v2.py
